requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
motor==3.3.1
pymongo[srv]==4.5.0
pydantic==2.12.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.2.1
python-multipart==0.0.20
pillow==12.0.0
reportlab==4.4.6
stripe==14.1.0
starlette==0.37.2
cryptography==46.0.3
httpx==0.28.1
Jinja2==3.1.6
weasyprint==67.0
dnspython==2.8.0
